# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /guarantor/request:
    post:
      operationId: requestGuarantor
      tags: [guarantor]
      summary: Request another verified customer to act as a digital guarantor
      description: Looks the person up by national ID, checks they're verified and in good financial standing, then creates a pending request.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuarantorRequestInput"
      responses:
        "200":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuarantorRequestResult"
        "400":
          description: Invalid request (self-guarantee, unverified/over-leveraged guarantor)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Firebase token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No user found with that national ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guarantor/respond:
    post:
      operationId: respondToGuarantorRequest
      tags: [guarantor]
      summary: Approve or decline an incoming guarantor request
      description: Approving raises the relationship to awaiting_admin_review; the guarantee only activates once an admin makes the final decision.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuarantorRespondInput"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuarantorRespondResult"
        "401":
          description: Missing or invalid Firebase token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No pending relationship with this id naming you as guarantor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /guarantor/network:
    get:
      operationId: getGuarantorNetwork
      tags: [guarantor]
      summary: Get the signed-in user's digital guarantor network
      description: Returns every guarantor relationship where this user is the requester or the guarantor (a user may have several of each), plus their remaining capacity on both sides.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Guarantor network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuarantorNetwork"
        "401":
          description: Missing or invalid Firebase token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
...
    GuarantorRelationshipSummary:
      type: object
      required:
        - id
        - requester_uid
        - requester_name
        - guarantor_uid
        - guarantor_name
        - status
        - max_amount
      properties:
        id:
          type: string
        requester_uid:
          type: string
        requester_name:
          type: string
        guarantor_uid:
          type: string
        guarantor_name:
          type: string
        status:
          type: string
          enum: [pending, awaiting_admin_review, approved, declined, rejected]
        max_amount:
          type: number
        requested_at:
          type: ["string", "null"]
          format: date-time
        responded_at:
          type: ["string", "null"]
          format: date-time
        admin_decision_reason:
          type: ["string", "null"]
    GuarantorCapacity:
      type: object
      required:
        - used_count
        - max_count
        - used_amount
        - max_amount
      properties:
        used_count:
          type: integer
        max_count:
          type: integer
        used_amount:
          type: number
        max_amount:
          type: number
    GuarantorNetwork:
      type: object
      required:
        - outgoing
        - incoming
        - guarantor_capacity
        - requester_capacity
      properties:
        outgoing:
          type: array
          items:
            $ref: "#/components/schemas/GuarantorRelationshipSummary"
        incoming:
          type: array
          items:
            $ref: "#/components/schemas/GuarantorRelationshipSummary"
        guarantor_capacity:
          $ref: "#/components/schemas/GuarantorCapacity"
        requester_capacity:
          $ref: "#/components/schemas/GuarantorCapacity"
    GuarantorRequestInput:
      type: object
      required:
        - guarantor_national_id
      properties:
        guarantor_national_id:
          type: string
    GuarantorRequestResult:
      type: object
      required:
        - id
        - status
        - guarantor_uid
        - guarantor_name
        - max_amount
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending]
        guarantor_uid:
          type: string
        guarantor_name:
          type: string
        max_amount:
          type: number
    GuarantorRespondInput:
      type: object
      required:
        - relationship_id
        - approve
      properties:
        relationship_id:
          type: string
        approve:
          type: boolean
    GuarantorRespondResult:
      type: object
      required:
        - id
        - requester_uid
        - status
      properties:
        id:
          type: string
        requester_uid:
          type: string
        status:
          type: string
          enum: [awaiting_admin_review, declined]
...