# lib/api-zod/src/generated/types/guarantorCapacity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuarantorCapacity {
  used_count: number;
  max_count: number;
  used_amount: number;
  max_amount: number;
}