# lib/api-zod/src/generated/types/guarantorNetwork.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuarantorCapacity } from './guarantorCapacity';
import type { GuarantorRelationshipSummary } from './guarantorRelationshipSummary';

export interface GuarantorNetwork {
  outgoing: GuarantorRelationshipSummary[];
  incoming: GuarantorRelationshipSummary[];
  guarantor_capacity: GuarantorCapacity;
  requester_capacity: GuarantorCapacity;
}