# lib/api-zod/src/generated/types/guarantorRelationshipSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuarantorRelationshipSummaryStatus } from './guarantorRelationshipSummaryStatus';

export interface GuarantorRelationshipSummary {
  id: string;
  requester_uid: string;
  requester_name: string;
  guarantor_uid: string;
  guarantor_name: string;
  status: GuarantorRelationshipSummaryStatus;
  max_amount: number;
  /** @nullable */
  requested_at?: Date | null;
  /** @nullable */
  responded_at?: Date | null;
  /** @nullable */
  admin_decision_reason?: string | null;
}