# lib/api-zod/src/generated/types/guarantorRequestResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuarantorRequestResultStatus } from './guarantorRequestResultStatus';

export interface GuarantorRequestResult {
  id: string;
  status: GuarantorRequestResultStatus;
  guarantor_uid: string;
  guarantor_name: string;
  max_amount: number;
}