# lib/api-zod/src/generated/types/guarantorRespondResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GuarantorRespondResultStatus } from './guarantorRespondResultStatus';

export interface GuarantorRespondResult {
  id: string;
  requester_uid: string;
  status: GuarantorRespondResultStatus;
}